pyproject.toml
[project]
name = "austinbrian-github-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.2"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
]
